Core: Debugs removes
