+++ crates/jstz_proto/src/api/contract.rs
@@ -12,10 +12,9 @@ use jstz_core::{
     host_defined,
     kv::Transaction,
     native::JsNativeObject,
-    runtime::{self, with_global_host},
+    runtime::{self},
     value::IntoJs,
 };
-use tezos_smart_rollup::prelude::debug_msg;
 
 use crate::{
     context::account::{Account, Address, Amount},
@@ -104,8 +103,6 @@ impl Contract {
         request: &JsNativeObject<Request>,
         context: &mut Context<'_>,
     ) -> JsResult<JsValue> {
-        host_defined!(context, host_defined);
-        with_global_host(|hrt| debug_msg!(hrt, "Call 1. \n"));
         // 1. Get address from request
         let address = request
             .deref()
@@ -116,12 +113,8 @@ impl Contract {
                 JsError::from_native(JsNativeError::error().with_message("Invalid host"))
             })?;
 
-        with_global_host(|hrt| debug_msg!(hrt, "Call 2. \n"));
-
         // 2. Set the referer of the request to the current contract address
         headers::test_and_set_referrer(&request.deref(), &self.contract_address)?;
-
-        with_global_host(|hrt| debug_msg!(hrt, "Call 3. \n"));
 
         // 3. Load, init and run!
         let trace_data = host_defined
@@ -149,23 +142,15 @@ impl ContractApi {
         args: &[JsValue],
         context: &mut Context<'_>,
     ) -> JsResult<JsValue> {
-        with_global_host(|hrt| {
-            debug_msg!(hrt, "Now I am calling a contract. Woohoo. \n")
-        });
-        with_global_host(|hrt| debug_msg!(hrt, "Small step. \n"));
         host_defined!(context, host_defined);
-        with_global_host(|hrt| debug_msg!(hrt, "Larger step. \n"));
         let mut tx = host_defined
             .get_mut::<Transaction>()
             .expect("Curent transaction undefined");
-
-        with_global_host(|hrt| debug_msg!(hrt, "Fisrt step done. \n"));
 
         let contract = Contract::from_js_value(this)?;
         let request: JsNativeObject<Request> =
             args.get_or_undefined(0).clone().try_into()?;
 
-        with_global_host(|hrt| debug_msg!(hrt, "Second step done. \n"));
         contract.call(tx.deref_mut(), &request, context)
     }
 
@@ -174,10 +159,6 @@ impl ContractApi {
         args: &[JsValue],
         context: &mut Context<'_>,
     ) -> JsResult<JsValue> {
-        println!("wooohooo I am creating a contract.");
-        with_global_host(|hrt| {
-            debug_msg!(hrt, "Wooohooo I am createing a contract (GH). \n")
-        });
         host_defined!(context, host_defined);
         let mut tx = host_defined.get_mut::<Transaction>().unwrap();
 
@@ -214,8 +195,6 @@ impl ContractApi {
             },
             context,
         )?;
-
-        with_global_host(|hrt| debug_msg!(hrt, "Yup its done. \n"));
 
         Ok(promise.into())
     }
