+++ crates/jstz_proto/src/context/account.rs
@@ -8,10 +8,7 @@ use jstz_core::{
 use jstz_crypto::public_key_hash::PublicKeyHash;
 
 use serde::{Deserialize, Serialize};
-use tezos_smart_rollup::{
-    prelude::debug_msg,
-    storage::path::{self, OwnedPath, RefPath},
-};
+use tezos_smart_rollup::storage::path::{self, OwnedPath, RefPath};
 
 pub type Address = PublicKeyHash;
 
@@ -60,11 +57,7 @@ impl Account {
     where
         'a: 'b,
     {
-        debug_msg!(hrt, "Get mut.");
-        //with_global_host(|hrt| debug_msg!(hrt, "Get mut 1. \n"));
         let account_entry = tx.entry::<Account>(hrt, Self::path(addr)?)?;
-        //with_global_host(|hrt| debug_msg!(hrt, "Get mut 2. \n"));
-        //with_global_host(|rt| debug_msg!(rt, "Woohoo I am here in get_mut."));
         Ok(account_entry.or_insert_default())
     }
 
@@ -101,11 +94,7 @@ impl Account {
         tx: &'a mut Transaction,
         addr: &Address,
     ) -> Result<Option<&'a mut String>> {
-        //with_global_host(|rt| debug_msg!(rt, "I am in contract code."));
-        debug_msg!(hrt, "Contract code 1. \n");
         let account = Self::get_mut(hrt, tx, addr)?;
-        debug_msg!(hrt, "Contract code 2. \n");
-        //with_global_host(|rt| debug_msg!(rt, "I have finished contract code."));
         Ok(account.contract_code.as_mut())
     }
 
@@ -215,8 +204,6 @@ mod test {
 
         // Act
         let amt = {
-            // This mutable borrow ends at the end of this block
-            //Account::balance(hrt, tx, &pkh).expect("Could not get balance")
             tx.entry::<Account>(hrt, Account::path(&pkh)?)?
                 .or_insert_default()
                 .amount
