+++ crates/jstz_proto/src/executor/contract.rs
@@ -117,14 +117,10 @@ impl Script {
         contract_address: &Address,
         context: &mut Context<'_>,
     ) -> Result<Self> {
-        with_global_host(|hrt| debug_msg!(hrt, "Loading 1. \n"));
         let src = with_global_host(|hrt| {
             Account::contract_code(hrt, tx, contract_address)?
                 .ok_or(Error::InvalidAddress)
         })?;
-        with_global_host(|hrt| debug_msg!(hrt, "Loading 2. \n"));
-        with_global_host(|hrt| debug_msg!(hrt, "Loading contract: {src:?}\n"));
-        //with_global_host(|hrt| debug_msg!(hrt, "Evaluating: {src:?}\n"));
 
         Ok(Self::parse(Source::from_bytes(&src), context)?)
     }
@@ -133,9 +129,7 @@ impl Script {
         src: Source<'_, R>,
         context: &mut Context<'_>,
     ) -> JsResult<Self> {
-        with_global_host(|hrt| debug_msg!(hrt, "Parse 1. \n"));
         let module = Module::parse(src, Some(Realm::new(context)?), context)?;
-        with_global_host(|hrt| debug_msg!(hrt, "Parse 2. \n"));
         Ok(Self(module))
     }
 
@@ -283,16 +277,13 @@ impl Script {
         request: &JsValue,
         context: &mut Context<'_>,
     ) -> JsResult<JsValue> {
-        with_global_host(|hrt| debug_msg!(hrt, "Script 1. \n"));
         // 1. Load script
         let script = Script::load(tx, &contract_address, context)?;
 
-        with_global_host(|hrt| debug_msg!(hrt, "Script 2. \n"));
         // 2. Evaluate the script's module
         let script_promise =
             script.init(contract_address.clone(), &operation_hash, context)?;
 
-        with_global_host(|hrt| debug_msg!(hrt, "Script 3. \n"));
         // 3. Once evaluated, call the script's handler
         let result = script_promise.then(
             Some(
@@ -312,8 +303,6 @@ impl Script {
             None,
             context,
         )?;
-
-        with_global_host(|hrt| debug_msg!(hrt, "Script 4. \n"));
 
         Ok(result.into())
     }
